Move BaseRestViewMixin to 'rest' submodule
