+++ django-rgd-3d/rgd_3d/rest/get.py
@@ -1,5 +1,5 @@
 from rest_framework.generics import RetrieveAPIView
-from rgd.mixins import BaseRestViewMixin
+from rgd.rest.mixins import BaseRestViewMixin
 from rgd_3d import models, serializers
