+++ django-rgd-3d/tests/test_permissions.py
@@ -1,5 +1,5 @@
 from django.apps import apps
-from rgd.mixins import BaseRestViewMixin
+from rgd.rest.mixins import BaseRestViewMixin
 from rgd.views import PermissionDetailView, PermissionListView, PermissionTemplateView
 from rgd_3d.urls import urlpatterns
 from rgd_testing_utils.helpers import check_model_permissions
