+++ django-rgd/rgd/rest/get.py
@@ -1,6 +1,6 @@
 from rest_framework.generics import RetrieveAPIView
 from rgd import models, serializers
-from rgd.mixins import BaseRestViewMixin
+from rgd.rest.mixins import BaseRestViewMixin
 
 
 class GetCollection(BaseRestViewMixin, RetrieveAPIView):
